% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{ints}
\alias{ints}
\title{This function produces a matrix which each row contains the first and last indexes
 to split a time series of length "n"}
\usage{
ints(n, l, p = 0, eq = TRUE)
}
\description{
n  = time series length
l  = interval length
p  = overlapping percentage
eq = last interval has length "l"
}
\keyword{internal}
