% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{knotLoc}
\alias{knotLoc}
\title{This function produces the knot location}
\usage{
knotLoc(data, k, degree, eqSpaced = FALSE)
}
\description{
data     = data
k        = number of B-spline densities
degree   = positive integer specifying the degree of the B-spline densities (default is 3)
eqSpaced = if TRUE, it produces equally spaced knots, otherwise it allocates the knots according to the periodogram
}
\keyword{internal}
