% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logplus}
\alias{logplus}
\title{C++ function for calculating the sum of the log of two log-values,i.e., logplus(log(a),log(b))=log(a+b).}
\usage{
logplus(x, y)
}
\description{
C++ function for calculating the sum of the log of two log-values,i.e., logplus(log(a),log(b))=log(a+b).
}
\keyword{internal}
