% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{lpost}
\alias{lpost}
\title{Unnormalised log posterior}
\usage{
lpost(omega, FZ, k, v, tau, tau.alpha, tau.beta, phi, phi.alpha, phi.beta,
  delta, delta.alpha, delta.beta, P, pdgrm, degree, db.list)
}
\description{
Unnormalised log posterior
}
\keyword{internal}
