% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{lprior}
\alias{lprior}
\title{log-prior}
\usage{
lprior(k, v, tau, tau.alpha, tau.beta, phi, phi.alpha, phi.beta, delta,
  delta.alpha, delta.beta, P)
}
\description{
log-prior
}
\keyword{internal}
