% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{qpsd}
\alias{qpsd}
\title{Compute unnormalised PSD using random mixture of B-splines}
\usage{
qpsd(omega, k, w, degree, db.list)

qpsd(omega, k, w, degree, db.list)
}
\description{
Compute unnormalised PSD using random mixture of B-splines

This function can deal with Inf values when taking exponential of weigths,
  unlike the function "qpsd" defined above
}
\keyword{internal}
