% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{diffMatrix}
\alias{diffMatrix}
\title{Difference Matrix. It works for second & third order penalties}
\usage{
diffMatrix(k, d = 2)
}
\description{
Difference Matrix. It works for second & third order penalties
}
\keyword{internal}
