% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_pspline.R
\name{spec_pspline}
\alias{spec_pspline}
\title{Spectrum estimate via p-splines}
\usage{
spec_pspline(
  data,
  l,
  p,
  eq = TRUE,
  Ntotal1,
  burnin1,
  thin1,
  Ntotal,
  burnin,
  thin,
  tau.alpha = 0.001,
  tau.beta = 0.001,
  phi.alpha = 1,
  phi.beta = 1,
  delta.alpha = 1e-04,
  delta.beta = 1e-04,
  k = NULL,
  eqSpacedKnots = FALSE,
  degree = 3,
  diffMatrixOrder = 1,
  printIter = 1000,
  recycl = FALSE,
  likePlot = FALSE
)
}
\arguments{
\item{data}{numeric vector}

\item{l}{length of the data subset to be analysed}

\item{p}{overlapping percentage of the data subsets}

\item{eq}{if is \code{TRUE}, last data subset has length \code{l}, even though the percentage is not \code{p}}

\item{Ntotal1}{total number of iterations to run the pilot Markov chain}

\item{burnin1}{number of initial iterations to be discarded in the pilot mcmc run}

\item{thin1}{thinning number in the pilot mcmc run (post-processing)}

\item{Ntotal}{total number of iterations to run the Markov chain}

\item{burnin}{number of initial iterations to be discarded}

\item{thin}{thinning number (post-processing)}

\item{tau.alpha, tau.beta}{prior parameters for tau (Inverse-Gamma)}

\item{phi.alpha, phi.beta}{prior parameters for phi (Gamma)}

\item{delta.alpha, delta.beta}{prior parameters for delta (Gamma)}

\item{k}{number of B-spline densities in the mixture}

\item{eqSpacedKnots}{logical value indicating whether the knots are equally spaced or defined according to the periodogram}

\item{degree}{positive integer specifying the degree of the B-spline densities (default is 3)}

\item{diffMatrixOrder}{positive integer specifying the order of the difference penalty matrix in the P-splines (default is 2)}

\item{printIter}{positive integer specifying the periodicity of the iteration number to be printed on screen (default 100)}

\item{recycl}{if \code{TRUE} the mcmc analysis for a data subset is used to calibrate the proposal for the succesor data subset.  It works for large \code{p} values.}

\item{likePlot}{if \code{TRUE}, a likelihood traceplot is generated for each data subset}
}
\value{
A list with S3 class `psds' containing the the power spectral estimates for the data subsets and and a list with relevant information about the analysis
}
\description{
This function allows to estimate the spectrum
}
\examples{
\dontrun{

set.seed(1)

# Generate AR(1) data with rho = 0.9
n = 128
data = arima.sim(n, model = list(ar = 0.9));
data = data - mean(data);

# Spectrum estimate via p-splines (may take some time)
spec = spec_pspline(data, l=50, p=90, Ntotal1=2000, burnin1=500, thin1=5,
                    Ntotal=1000, burnin=500, thin=5, k=30, recycl = TRUE);

image(spec) # Plot log PSD (see documentation of image.plot.psd)

}
}
\seealso{
\link{gibbs_pspline}
}
